21363: Unlist libcurl4 as a gems dependency
In fact, it still is a dependency. However, the package is named
`libcurl4t64` after Debian's 64-bit time transition. The right package
is automatically pulled in when we install `curl` which is also listed
as a dependency. Letting the dependency be implicit is the most
straightforward way to support distros with either name.

Arvados-DCO-1.1-Signed-off-by: Brett Smith <brett.smith@curii.com>